requirements.txt
Django==5.0.4
Pillow==10.2.0
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-tailwind==1.0.0
django-tailwind==3.8.0